Sept 26 revise

+++ Working/LiveHacks/Livehack2.py
@@ -3,5 +3,5 @@ print("""
 ┃This program will convert hours to days and hours.┃
 ----------------------------------------------------
 """)
-hours = int(input("Input hours here: "))
+hours = float(input("Input hours here: "))
 print((hours//24), "days,", (hours % 24), "hours.")
